requirements.txt
aiogram==3.10.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
aiohttp==3.9.1
apscheduler==3.10.4